+++ core/blocks.py
@@ -17,8 +17,11 @@ class VideoBlock(blocks.StructBlock):
     height = blocks.IntegerBlock()
     video = MediaChooserBlock()
 
+    class Meta:
+        icon = 'fa-play'
 
-class ObjectiveItem(blocks.StructBlock):
+
+class Item(blocks.StructBlock):
     item = blocks.CharBlock(max_length=255)
 
 
@@ -81,6 +84,15 @@ class ImageBlock(ImageChooserBlock):
         help_text = 'Include an image'
         template = 'core/includes/_image.html'
         icon = 'image'
+
+
+class SimpleVideoBlock(blocks.StructBlock):
+    video = MediaChooserBlock()
+
+    class Meta:
+        help_text = 'Include a video'
+        template = 'core/includes/_hero_video.html'
+        icon = 'fa-play'
 
 
 class ButtonBlock(blocks.StructBlock):
@@ -184,3 +196,11 @@ class StepByStepBlock(blocks.StructBlock):
 
     class Meta:
         template = 'learn/step_by_step.html'
+
+
+class ITAQuoteBlock(blocks.StructBlock):
+    quote = blocks.RichTextBlock()
+    author = blocks.CharBlock(max_length=255)
+
+    class Meta:
+        template = 'learn/ita_quote.html'