+++ tests/conftest.py
@@ -224,6 +224,14 @@ def mock_get_company_profile(patch_get_company_profile):
 
 
 @pytest.fixture
+def patch_get_user_lesson_completed():
+    yield mock.patch(
+        'directory_sso_api_client.sso_api_client.user.get_user_lesson_completed',
+        return_value=create_response(status_code=200, json_body={'result': 'ok'})
+    ).start()
+
+
+@pytest.fixture
 def patch_update_company_profile():
     yield mock.patch('core.helpers.update_company_profile', return_value=None)
 
@@ -271,6 +279,14 @@ def patch_set_user_page_view():
     yield mock.patch(
         'directory_sso_api_client.sso_api_client.user.set_user_page_view',
         return_value=create_response(status_code=200, json_body={'result': 'ok'})
+    ).start()
+
+
+@pytest.fixture
+def patch_export_plan():
+    yield mock.patch(
+        'directory_api_client.api_client.exportplan.exportplan_list',
+        return_value=create_response(status_code=200, json_body=[{'id': 1}])
     ).start()