+++ tests/unit/core/test_wagtail_hooks.py
@@ -200,3 +200,44 @@ def test_estimated_read_time_calculation(rf, domestic_homepage):
         request=request
     )
     assert response > 0
+
+
+@pytest.mark.django_db
+def test_set_lesson_pages_topic_id(rf, topics_with_lessons):
+
+    request = rf.get('/')
+    request.user = AnonymousUser()
+
+    # Rest the topic_block_id for all lessons
+    curated_page = topics_with_lessons[0][0]
+    for topic in topics_with_lessons[0][1]:
+        topic.topic_block_id = None
+        topic.save()
+
+    response = wagtail_hooks.set_lesson_pages_topic_id(
+        page=curated_page,
+        request=request
+    )
+
+    for topic in response.topics:
+        for lesson in topic.value['pages']:
+            assert lesson.specific.topic_block_id == topic.id
+
+
+@pytest.mark.django_db
+def test_set_lesson_pages_topic_id_removed(rf, topics_with_lessons):
+
+    request = rf.get('/')
+    request.user = AnonymousUser()
+
+    curated_page = topics_with_lessons[0][0]
+    # Remove the second lesson from topic
+    topic_page_2 = curated_page.topics[0].value['pages'][1]
+    curated_page.topics[0].value['pages'].remove(topic_page_2)
+
+    wagtail_hooks.set_lesson_pages_topic_id(
+        page=curated_page,
+        request=request
+    )
+
+    assert topic_page_2.specific.topic_block_id is None